+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fix
+- Fix user not appearing in the user list if no activity has been recorded for the user yet (see [sprinkle-account#25](https://github.com/userfrosting/sprinkle-account/pull/25)).
+
 ## [6.0.0](https://github.com/userfrosting/sprinkle-account/compare/6.0.0-rc.5...6.0.0) - 2026-06-12
 - No changes.